+++ volmdlr/edges.py
@@ -328,6 +328,11 @@ class Edge(dc.DessiaObject):
                 break
         return distance
 
+    @property
+    def simplify(self):
+        """Search another simplified edge that can represent the edge."""
+        return self
+
 
 class Line(dc.DessiaObject):
     """
@@ -633,6 +638,8 @@ class LineSegment(Edge):
         :return: shared line segment section.
         """
         if self.__class__ != other_linesegment.__class__:
+            if self.__class__ == other_linesegment.simplify.__class__:
+                return self.get_shared_section(other_linesegment.simplify)
             return []
         if not self.direction_vector().is_colinear_to(other_linesegment.direction_vector()) or \
                 (not any(self.point_belongs(point, 1e-6)
@@ -775,6 +782,7 @@ class BSplineCurve(Edge):
         self.knot_multiplicities = knot_multiplicities
         self.weights = weights
         self.periodic = periodic
+        self._simplified = None
 
         points = [[*point] for point in control_points]
         if weights is None:
@@ -843,6 +851,36 @@ class BSplineCurve(Edge):
             knots=self.knots[::-1],
             weights=self.weights,
             periodic=self.periodic)
+
+    @property
+    def simplify(self):
+        """Search another simplified edge that can represent the bspline."""
+        if self._simplified is None:
+            class_sufix = self.__class__.__name__[-2:]
+            lineseg_class = getattr(sys.modules[__name__], 'LineSegment' + class_sufix)
+            lineseg = lineseg_class(self.points[0], self.points[-1])
+            if all(lineseg.point_belongs(pt) for pt in self.points):
+                self._simplified = lineseg
+                return lineseg
+            colinear = True
+            previous_vec = None
+            points_ = [self.points[0]]
+            for i, (point1, point2) in enumerate(zip(self.points[:-1], self.points[1:])):
+                vec = point2 - point1
+                if i == 0:
+                    previous_vec = vec
+                elif not vec.is_colinear_to(previous_vec):
+                    colinear = False
+                    points_.append(point2)
+                previous_vec = vec
+            if not colinear:
+                arc_class_ = getattr(sys.modules[__name__], 'Arc' + class_sufix)
+                try_arc = arc_class_(points_[0], points_[int(len(points_) / 2)], points_[-1])
+                if all(try_arc.point_belongs(point, 1e-6) for point in self.points):
+                    self._simplified = try_arc
+                    return try_arc
+            self._simplified = self
+        return self._simplified
 
     @classmethod
     def from_geomdl_curve(cls, curve, name: str = ""):
@@ -1203,30 +1241,6 @@ class BSplineCurve(Edge):
         :return: A B-spline curve from points interpolation
         :rtype: :class:`volmdlr.edges.BSplineCurve`
         """
-        # class_sufix = cls.__name__[-2:]
-        # lineseg_class = getattr(sys.modules[__name__], 'LineSegment' + class_sufix)
-        # lineseg = lineseg_class(points[0], points[-1])
-        # if all(lineseg.point_belongs(pt) for pt in points):
-        #     return lineseg
-        # colinear = True
-        # previous_vec = None
-        # points_ = [points[0]]
-        # for i, (point1, point2) in enumerate(zip(points[:-1], points[1:])):
-        #     vec = point2 - point1
-        #     if i == 0:
-        #         previous_vec = vec
-        #     elif not vec.is_colinear_to(previous_vec):
-        #         colinear = False
-        #         points_.append(point2)
-        #     previous_vec = vec
-        # if not colinear:
-        #     arc_class_ = getattr(sys.modules[__name__], 'Arc' + class_sufix)
-        #     try:
-        #         try_arc = arc_class_(points_[0], points_[1], points_[2])
-        #     except Exception:
-        #         print(True)
-        #     if all(try_arc.point_belongs(point, 1e-6) for point in points):
-        #         return try_arc
         curve = bspline_fitting.interpolate_curve([[*point] for point in points], degree, centripetal=True)
 
         bsplinecurve = cls.from_geomdl_curve(curve, name=name)
@@ -1393,6 +1407,8 @@ class BSplineCurve(Edge):
         :return: shared arc section.
         """
         if self.__class__ != other_bspline2.__class__:
+            if self.simplify.__class__ == other_bspline2.__class__:
+                return self.simplify.get_shared_section(other_bspline2)
             return []
         if self.__class__.__name__[-2:] == '3D':
             if self.bounding_box.distance_to_bbox(other_bspline2.bounding_box) > 1e-7:
@@ -2702,6 +2718,8 @@ class Arc(Edge):
         :return: shared arc section.
         """
         if self.__class__ != other_arc2.__class__:
+            if self.__class__ == other_arc2.simplify.__class__:
+                return self.get_shared_section(other_arc2.simplify)
             return []
         if not self.center.is_close(other_arc2.center) or self.radius != self.radius or \
                 not any(self.point_belongs(point) for point in [other_arc2.start,