+++ volmdlr/faces.py
@@ -4641,12 +4641,6 @@ class BSplineSurface3D(Surface3D):
 
                 if self.y_periodicity:
                     points = self._repair_periodic_boundary_points(bspline_curve3d, points, 'y')
-                # points_ = [points[0]]
-                # for point in points[1:]:
-                #     if not point.is_close(points[-1]):
-                #         points_.append(point)
-                # if len(points_) < 2:
-                #     return []
 
                 return [vme.BSplineCurve2D.from_points_interpolation(
                     points=points, degree=bspline_curve3d.degree, periodic=bspline_curve3d.periodic)]